backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
groq>=0.4.0
django-environ>=0.11.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
gunicorn
whitenoise
dj-database-url
psycopg2-binary
